+++ src/interpreter.rs
@@ -1,22 +1,17 @@
-use crate::binary::{read_i16, read_u16, read_u8};
-use crate::instruction_set::Instruction;
-use crate::interpreter::Result::Continue;
-use crate::interpreter::Result::Return;
-use crate::thread::Frame;
 use std::fs::read;
 use std::io::Cursor;
 use std::rc::Rc;
 
-enum Result {
-    Return(u64),
-    Continue,
-}
+use crate::binary::{read_i16, read_u16, read_u8};
+use crate::instruction::instruction;
+use crate::instruction_set::Instruction;
+use crate::thread::Frame;
 
-// TODO: 1. sprit instruction code to function
-// TODO: 2. create return type for instruction, e.g. CONTINUE/RETURN/INVOKE
+use crate::instruction::Result::Return;
+use crate::JVM;
+
 pub fn interpret(frame: &mut Frame) -> u64 {
     let code = &frame.current_method.get_code_attribute().code;
-
     let cursor = &mut Cursor::new(code.as_slice());
     let local_variable = &mut frame.local_variable;
     let operand_stack = &mut frame.operand_stack;
@@ -24,113 +19,13 @@ pub fn interpret(frame: &mut Frame) -> u64 {
     let result = loop {
         frame.pc = cursor.position();
         let instruction_code = read_u8(cursor);
-        let instruction = Instruction::from(instruction_code);
         // println!(
         //     "[DEBUG] -- frame.pc: {} instruction: {:#?}(0x{:x})",
         //     frame.pc, instruction, instruction_code
         // );
-        let result = match instruction {
-            Instruction::ICONST_0 => {
-                operand_stack.push(0);
-                Continue
-            }
-            Instruction::ICONST_1 => {
-                operand_stack.push(1);
-                Continue
-            }
-            Instruction::ICONST_2 => {
-                operand_stack.push(2);
-                Continue
-            }
-
-            Instruction::ISTORE_0 => {
-                let val = operand_stack.pop().unwrap();
-                local_variable[0] = val;
-                Continue
-            }
-            Instruction::ISTORE_1 => {
-                let val = operand_stack.pop().unwrap();
-                local_variable[1] = val;
-                Continue
-            }
-            Instruction::ISTORE_2 => {
-                let val = operand_stack.pop().unwrap();
-                local_variable[2] = val;
-                Continue
-            }
-
-            Instruction::ILOAD_0 => {
-                let val = local_variable[0];
-                operand_stack.push(val);
-                Continue
-            }
-            Instruction::ILOAD_1 => {
-                let val = local_variable[1];
-                operand_stack.push(val);
-                Continue
-            }
-            Instruction::ILOAD_2 => {
-                let val = local_variable[2];
-                operand_stack.push(val);
-                Continue
-            }
-
-            Instruction::IADD => {
-                let val2 = operand_stack.pop().unwrap();
-                let val1 = operand_stack.pop().unwrap();
-
-                let result = val1 + val2;
-                operand_stack.push(result);
-                Continue
-            }
-
-            Instruction::SIPUSH => {
-                // TODO: handle type...
-                let val = read_u16(cursor);
-                operand_stack.push(val as u64);
-                Continue
-            }
-
-            Instruction::IINC => {
-                let index = read_u8(cursor);
-                let const_val = read_u8(cursor);
-
-                local_variable[index as usize] += const_val as u64;
-                Continue
-            }
-
-            Instruction::IF_ICMPGE => {
-                let current_pc = cursor.position() - 1;
-                let next_pc_offset = read_i16(cursor);
-                let val2 = operand_stack.pop().unwrap();
-                let val1 = operand_stack.pop().unwrap();
-
-                if val1 >= val2 {
-                    cursor.set_position((current_pc as i64 + next_pc_offset as i64) as u64);
-                }
-                Continue
-            }
-
-            Instruction::GOTO => {
-                // TODO! unsigned offset calc is suck
-                let current_pc = cursor.position() - 1;
-                let next_pc_offset = read_i16(cursor);
-                cursor.set_position((current_pc as i64 + next_pc_offset as i64) as u64);
-                Continue
-            }
-
-            Instruction::IRETURN => {
-                let val = operand_stack.pop().unwrap();
-                Return(val)
-            }
-            _ => panic!(
-                "Instruction {:#?}(0x{:x}) isn't implemented yet",
-                instruction, instruction_code
-            ),
-        };
-        match result {
-            Continue => {}
+        match instruction(cursor, instruction_code, local_variable, operand_stack) {
             Return(v) => break v,
+            _ => {}
         }
     };
 
@@ -170,10 +65,11 @@ fn test_invoke_loop() {
 }
 
 #[cfg(test)]
-mod frame_test {
+pub mod frame_test {
+    use std::collections::HashMap;
+
     use crate::class::Class;
     use crate::class_attributes::{AttributeInfo, CodeAttributeInfo, MethodInfo};
-    use std::collections::HashMap;
 
     pub fn dummy_class() -> Class {
         Class {
